# bpo-45923: Add `RESUME_QUICK`