--- a/Misc/NEWS.d/next/Core and Builtins/2022-02-09-20-21-43.bpo-45923.tJ4gDX.rst
+++ b/Misc/NEWS.d/next/Core and Builtins/2022-02-09-20-21-43.bpo-45923.tJ4gDX.rst
@@ -1,0 +1,1 @@
+Add a quickened form of :opcode:`RESUME` that skips quickening checks.