--- a/Python/ceval.c
+++ b/Python/ceval.c
@@ -1785,9 +1785,6 @@ handle_eval_breaker:
         }
 
         TARGET(RESUME) {
-            assert(tstate->cframe == &cframe);
-            assert(frame == cframe.current_frame);
-
             int err = _Py_IncrementCountAndMaybeQuicken(frame->f_code);
             if (err) {
                 if (err < 0) {
@@ -1798,6 +1795,13 @@ handle_eval_breaker:
                 first_instr = frame->f_code->co_firstinstr;
                 next_instr = first_instr + nexti;
             }
+            JUMP_TO_INSTRUCTION(RESUME_QUICK);
+        }
+
+        TARGET(RESUME_QUICK) {
+            PREDICTED(RESUME_QUICK);
+            assert(tstate->cframe == &cframe);
+            assert(frame == cframe.current_frame);
             frame->f_state = FRAME_EXECUTING;
             if (_Py_atomic_load_relaxed(eval_breaker) && oparg < 2) {
                 goto handle_eval_breaker;
@@ -2265,7 +2269,7 @@ handle_eval_breaker:
             Py_DECREF(v);
             if (err != 0)
                 goto error;
-            PREDICT(JUMP_ABSOLUTE);
+            PREDICT(JUMP_ABSOLUTE_QUICK);
             DISPATCH();
         }
 
@@ -2277,7 +2281,7 @@ handle_eval_breaker:
             Py_DECREF(v);
             if (err != 0)
                 goto error;
-            PREDICT(JUMP_ABSOLUTE);
+            PREDICT(JUMP_ABSOLUTE_QUICK);
             DISPATCH();
         }
 
@@ -3420,7 +3424,7 @@ handle_eval_breaker:
             if (_PyDict_SetItem_Take2((PyDictObject *)map, key, value) != 0) {
                 goto error;
             }
-            PREDICT(JUMP_ABSOLUTE);
+            PREDICT(JUMP_ABSOLUTE_QUICK);
             DISPATCH();
         }
 
@@ -4057,8 +4061,6 @@ handle_eval_breaker:
         }
 
         TARGET(JUMP_ABSOLUTE) {
-            PREDICTED(JUMP_ABSOLUTE);
-            assert(oparg < INSTR_OFFSET());
             int err = _Py_IncrementCountAndMaybeQuicken(frame->f_code);
             if (err) {
                 if (err < 0) {
@@ -4069,9 +4071,7 @@ handle_eval_breaker:
                 first_instr = frame->f_code->co_firstinstr;
                 next_instr = first_instr + nexti;
             }
-            JUMPTO(oparg);
-            CHECK_EVAL_BREAKER();
-            DISPATCH();
+            JUMP_TO_INSTRUCTION(JUMP_ABSOLUTE_QUICK);
         }
 
         TARGET(JUMP_NO_INTERRUPT) {
@@ -4086,6 +4086,7 @@ handle_eval_breaker:
         }
 
         TARGET(JUMP_ABSOLUTE_QUICK) {
+            PREDICTED(JUMP_ABSOLUTE_QUICK);
             assert(oparg < INSTR_OFFSET());
             JUMPTO(oparg);
             CHECK_EVAL_BREAKER();
@@ -4546,7 +4547,6 @@ handle_eval_breaker:
 
         TARGET(KW_NAMES) {
             assert(call_shape.kwnames == NULL);
-            assert(oparg < PyTuple_GET_SIZE(consts));
             call_shape.kwnames = GETITEM(consts, oparg);
             DISPATCH();
         }