@markshannon 2022-02-10
I note that https://github.com/python/cpython/pull/31240 also shows a ~1% speedup.
I wonder, is this a slow common base build, or two independent speedups?