@brandtbucher 2022-02-10
> I note that #31240 also shows a ~1% speedup. I wonder, is this a slow common base build, or two independent speedups?

I'm pretty sure they're independent, since even though the `main` base changeset is the same, they use two different builds and two different `pyperformance` runs.

All four combinations of PR branches and `main` runs show a 1% improvement:

- `resume-quick` vs `main` (from this PR)
- `resume-quick` vs `main` (from #31240)
- `unpack-sequence` vs `main` (from this PR)
- `unpack-sequence` vs `main` (from #31240)

I know the UK benchmarking machine is a bit more stable than the US one, though, so if you want to verify those numbers I won't be offended. 🙃