@markshannon 2022-02-10
OK, merging this. We can quicken `RESUME (oparg <2)` to `NOP` once `f_state` has been removed.